refactor : BaseTimeEntity생성 및 id들 다 int=> Long으로 변경

+++ src/main/java/com/example/instaclone/controller/UserController.java
@@ -20,14 +20,14 @@ public class UserController {
     private final UserService userService;
 
     @GetMapping("/user/{page.UserId}")
-    public String profile(@PathVariable int pageUserId, Model model, @AuthenticationPrincipal PrincipalDetails principalDetails) {
+    public String profile(@PathVariable Long pageUserId, Model model, @AuthenticationPrincipal PrincipalDetails principalDetails) {
         UserProfileDto dto = userService.userProfile(pageUserId,principalDetails.getUser().getId());
         model.addAttribute("user",dto);
         return "user/profile";
     }
 
     @GetMapping("/user/{id}/update")
-    public String update(@PathVariable int id, @AuthenticationPrincipal PrincipalDetails principalDetails){
+    public String update(@PathVariable Long id, @AuthenticationPrincipal PrincipalDetails principalDetails){
 
         Authentication auth = SecurityContextHolder.getContext().getAuthentication();
         PrincipalDetails mPrincipalDetails = (PrincipalDetails) auth.getPrincipal();
