+++ src/main/java/com/example/instaclone/controller/api/SubscribeApiController.java
@@ -1,7 +1,6 @@
 package com.example.instaclone.controller.api;
 
 import com.example.instaclone.config.auth.PrincipalDetails;
-import com.example.instaclone.dto.CommonResponseDto;
 import com.example.instaclone.service.SubscribeService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -19,13 +18,13 @@ public class SubscribeApiController {
     private final SubscribeService subscribeService;
 
     @PostMapping("/api/subscribe/{userId}")
-    public ResponseEntity<?> subscribe(@AuthenticationPrincipal PrincipalDetails principalDetails, @PathVariable int toUserId) {
+    public ResponseEntity<?> subscribe(@AuthenticationPrincipal PrincipalDetails principalDetails, @PathVariable Long toUserId) {
         subscribeService.subscribe(principalDetails.getUser().getId(),toUserId);
         return new ResponseEntity<>(new CommonResponseDto<>(1,"구독 성공",null), HttpStatus.OK);
     }
 
     @DeleteMapping("/api/subscribe/{userId}")
-    public ResponseEntity<?> unsubscribe(@AuthenticationPrincipal PrincipalDetails principalDetails, @PathVariable int toUserId) {
+    public ResponseEntity<?> unsubscribe(@AuthenticationPrincipal PrincipalDetails principalDetails, @PathVariable Long toUserId) {
         subscribeService.unsubscribe(principalDetails.getUser().getId(),toUserId);
         return new ResponseEntity<>(new CommonResponseDto<>(1,"구독 취소",null), HttpStatus.OK);
     }
